Removed dependeny on remoting that was used to couple TranformManager to CsBase by inverting the dependency.
